# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Inventario API",
    "description": "REST API for Inventario — a multi-tenant, white-label asset management platform for sports federations, municipalities, schools, and clubs. Open-source under EUPL-1.2.",
    "version": "0.1.0",
    "contact": {
      "name": "Inventario · LTK Solutions",
      "url": "https://inventario.sportup.sk",
      "email": "inventario@sportup.sk"
    },
    "license": {
      "name": "EUPL-1.2",
      "url": "https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12"
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Microsoft Entra ID access token (v2.0). Multi-tenant: the JWT `tid` claim resolves the Organisation; the `oid` claim resolves the User. See apps/api/README.md for the device code flow."
      }
    },
...
  },
  "paths": {
...
    "/v1/categories": {
      "get": {
        "summary": "List categories",
        "tags": [
          "Categories"
        ],
        "description": "Returns a paginated list of categories, sorted by sortOrder then name. Soft-deleted categories are excluded. Optional filters: parentId, assetType, isActive.",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "in": "query",
            "name": "skip",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^([a-f\\d]{24}|null)$"
            },
            "in": "query",
            "name": "parentId",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "assetType",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false",
                "1",
                "0"
              ]
            },
            "in": "query",
            "name": "isActive",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": {}
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "limit": {
                          "type": "integer",
                          "exclusiveMinimum": true,
                          "minimum": 0
                        },
                        "skip": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "hasMore": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "total",
                        "limit",
                        "skip",
                        "hasMore"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data",
                    "pagination"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new category",
        "tags": [
          "Categories"
        ],
        "description": "Creates a new category. Slug must be unique. If parentId is supplied, the parent must exist. Requires ASSET_MANAGER or ADMIN role.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "slug": {
                    "type": "string",
                    "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                    "maxLength": 200
                  },
                  "parentId": {
                    "type": "string",
                    "pattern": "^[a-f\\d]{24}$",
                    "nullable": true,
                    "default": null
                  },
                  "assetType": {
                    "type": "string",
                    "enum": [
                      "IT",
                      "SPORTS_GEAR",
                      "TRAINING_EQUIPMENT",
                      "OFFICE_EQUIPMENT",
                      "MEDIA",
                      "COMMUNICATION",
                      "OTHER"
                    ]
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "nullable": true,
                    "default": null
                  },
                  "icon": {
                    "type": "string",
                    "maxLength": 50,
                    "nullable": true,
                    "default": null
                  },
                  "color": {
                    "type": "string",
                    "pattern": "^#[0-9a-fA-F]{6}$",
                    "nullable": true,
                    "default": null
                  },
                  "approverIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "pattern": "^[a-f\\d]{24}$"
                    },
                    "default": []
                  },
                  "requiresApprovalByDefault": {
                    "type": "boolean",
                    "default": true
                  },
                  "maxLoanDays": {
                    "type": "integer",
                    "exclusiveMinimum": true,
                    "minimum": 0,
                    "maximum": 3650,
                    "nullable": true,
                    "default": null
                  },
                  "isActive": {
                    "type": "boolean",
                    "default": true
                  },
                  "sortOrder": {
                    "type": "integer",
                    "default": 0
                  }
                },
                "required": [
                  "name",
                  "assetType"
                ],
                "additionalProperties": false,
                "description": "Telo pre vytvorenie kategórie; slug je voliteľný (server odvodí z name)."
              }
            }
          },
          "description": "Telo pre vytvorenie kategórie; slug je voliteľný (server odvodí z name)."
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
    "/v1/categories/{id}": {
      "get": {
        "summary": "Get a single category by ID",
        "tags": [
          "Categories"
        ],
        "description": "Returns one category by its _id. 404 if not found or soft-deleted.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[a-f\\d]{24}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update an existing category",
        "tags": [
          "Categories"
        ],
        "description": "Partial update — only provided fields are changed. Slug uniqueness and parent existence are revalidated on change. Records an audit event with a per-field diff. Requires ASSET_MANAGER or ADMIN role.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "slug": {
                    "type": "string",
                    "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                    "maxLength": 200
                  },
                  "parentId": {
                    "type": "string",
                    "pattern": "^[a-f\\d]{24}$",
                    "description": "MongoDB ObjectId (24 hex znakov)",
                    "nullable": true,
                    "default": null
                  },
                  "assetType": {
                    "type": "string",
                    "enum": [
                      "IT",
                      "SPORTS_GEAR",
                      "TRAINING_EQUIPMENT",
                      "OFFICE_EQUIPMENT",
                      "MEDIA",
                      "COMMUNICATION",
                      "OTHER"
                    ]
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "nullable": true,
                    "default": null
                  },
                  "icon": {
                    "type": "string",
                    "maxLength": 50,
                    "nullable": true,
                    "default": null
                  },
                  "color": {
                    "type": "string",
                    "pattern": "^#[0-9a-fA-F]{6}$",
                    "nullable": true,
                    "default": null
                  },
                  "approverIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "pattern": "^[a-f\\d]{24}$",
                      "description": "MongoDB ObjectId (24 hex znakov)"
                    },
                    "default": []
                  },
                  "requiresApprovalByDefault": {
                    "type": "boolean",
                    "default": true
                  },
                  "maxLoanDays": {
                    "type": "integer",
                    "exclusiveMinimum": true,
                    "minimum": 0,
                    "maximum": 3650,
                    "nullable": true,
                    "default": null
                  },
                  "isActive": {
                    "type": "boolean",
                    "default": true
                  },
                  "sortOrder": {
                    "type": "integer",
                    "default": 0
                  }
                },
                "additionalProperties": false,
                "description": "Čiastočná aktualizácia kategórie; všetky polia voliteľné."
              }
            }
          },
          "description": "Čiastočná aktualizácia kategórie; všetky polia voliteľné."
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[a-f\\d]{24}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Soft-delete a category",
        "tags": [
          "Categories"
        ],
        "description": "Marks a category as deleted. Refuses deletion if the category has any non-deleted child categories (would orphan them), or if any non-deleted assets reference it. Requires ADMIN role.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[a-f\\d]{24}$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "enum": [
                    "null"
                  ],
                  "nullable": true
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local dev"
    },
    {
      "url": "https://api.inventario.sportup.sk",
      "description": "Production (planned Q3 2026)"
    }
  ],
...
}